fix: stabilize full-preset on RP2350 (64-bit bitmap, irq, proto sizing)
Several latent bugs surfaced once the full preset runs on real hardware:

- task_used_bitmap / exit_retain_bitmap were uint32 but KERNEL_MAX_TASKS=64,
  so IDs >= 32 silently collided. Widened to uint64 and added a compile-time
  guard. (stats.c, test_stats.c updated to match the new return type.)
- irq_register() also enables the NVIC line; on RP2350 the numeric IRQ range
  covers live UART/timer sources whose status the test stub does not clear,
  causing an interrupt storm that hangs the suite mid-way. test_isr_pool now
  registers under PRIMASK, clamps to BOARD_IRQ_COUNT, and unregisters on
  exit. BH-pool test expects one fewer slot when FAULT_ENDPOINT owns a BH.
- drv_msg_t / fs_msg_t were not padded to KERN_EP_MSG_SIZE (128 on full),
  so IPC could read/write past the message struct. Added reserved padding
  guarded by KERN_EP_MSG_SIZE.
- Promoted root_bootstrap's static set_initial_arg to the public
  task_set_initial_arg(task_id, arg) in task.h (reused by Phase 2 init).
- Added endpoint_msg_size() helper (IRQ-save guarded lookup).
- tools/openocd.cfg: adjusted for stable CMSIS-DAP bring-up on this board.

These are prerequisites for the Phase 2 fault-endpoint work that follows.

Tested on Pico 2 W: test suite now runs past the irq module without hanging.

+++ src/kernel/stats/stats.c
@@ -22,7 +22,7 @@ static uint32_t subsys_counters[STATS_SUBSYS_MAX][STATS_COUNTER_MAX];
  * 外部引用
  *============================================================================*/
 
-extern uint32_t task_get_used_bitmap(void);
+extern uint64_t task_get_used_bitmap(void);
 /*============================================================================
  * 初始化
  *============================================================================*/
@@ -68,9 +68,9 @@ void stats_tick_update(void) {
     }
     last_stat_tick = tick;
 
-    uint32_t bitmap = task_get_used_bitmap();
+    uint64_t bitmap = task_get_used_bitmap();
     for (int i = 0; i < KERNEL_MAX_TASKS; i++) {
-        if (bitmap & (1U << i)) {
+        if (bitmap & (1ULL << i)) {
             tcb_t *tcb = task_get_tcb((task_id_t)i);
             if (tcb) {
                 tcb->cpu_usage = tcb->total_ticks * 10000 / elapsed;
